pyproject.toml
[project]
name = "llm_project"  # or "llm-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Nevivaghani", email = "nevivaghani793@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain (>=0.3.22,<0.4.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "google-ai-generativelanguage (>=0.6.17,<0.7.0)",
    "langchain-google-genai (>=2.1.2,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "protobuf (>=5.0.0)",
    "streamlit (>=1.44.0,<2.0.0)",
    "google-auth (>=2.38.0,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "instructorembedding (==1.0.1)",
    "huggingface-hub (==0.15.1)",
    "sentence-transformers (==2.2.2)"

]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
langchain-community = "^0.3.20"
langchain-core = "^0.3.49"

